podoclinic/requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
celery>=5.3.6
redis>=5.0.1 